feat!: Added transitions to EntryPoints per #4.
Earlier EntryPoint directly defined the target state
they wanted the machine be in if the EntryPoint was used to
start the machine. It had the drawback of not being able to
execute actions when that EntryPoint was used.
This change introduces list of transition ids to be defined
for an EntryPoint. It will enable us executing actions when
the selected transition is happening.
Selecting the right transition happens the same way as it was
already defined state transitions (actually the same code is used).