fix: StateMachine now does not await on monitor notifications.
Earlier StateMachine awaited on monitor notifications. This resulted
that if a monitor (e.g. visual_monitor) could not complete its operation
during a notification it had the whole program suspended.
With this change StateMachine is not awaiting the notifications, hence
it is not suspended if notification takes time to complete.