feat: Added onError for transitions.
Transitions can have the minInterval defined that sets minimal interval
between two executions of the transition. Earlier only an exception
was thrown in case the interval between the execs were shorter than
minInterval. Using exception meant that only in the layer that invokes
the fire() operation you could catch the exception and initiate some
actions like firing another event. With the introduction of the onError
attribute to transitions one can define the onError handler that would
be called instead of throwing the exception. This way allowing to put
error handling right into the state machine definition instead of an
upper layer. Right now it is only the above mentioned minInterval check
that triggers this, but later it could be extended to general error
handling during state machine operations.