fix: Catch only Exceptions
During investigating issue #24 besides the immediate cause of that
problem (not defined transition) the root cause was identified as
improper exception handling: When the lack of the transition detected
in PlantUMLConverter an AssertionError is thrown. This is an Error and
not supposed to be caught, but in VisualMonitor._connectCord() it was
caught by try {...} catch (e) {...}.
This was replaced by catching only Exceptions (and not Errors).

During left-right look, some other places were identified where the
same problem had to be fixed. Few autotests were also fixed as they
were generating Errors that were before hidden by the above mentioned
catch issue.