fix!: commented out BottomSheetCreator
Since PersistentBottomSheetController type attribute was removed in
relation to https://github.com/flutter/flutter/issues/137424
the BottomSheetCreator class was commented out to have a quick answer
to that change. This way hisma_flutter is operational with
flutter versions with our without type parameter for
PersistentBottomSheetController. BottomSheetCreator is an experimental
class and its fix is postponed.