refactor!: monitor notification for all state entry
Earlier a single notification was sent to monitors in case of all state
changes that occurred as a result of a fire. For example if fire
triggered a state change and in the new state inside the onEntry callback
another fire took the machine yet into another state, only a single state
change notification was sent to monitors at the completion of the the
second state change. This resulted that state changes could not be followed
in details, hence the design is now changed to send notification at each
and every state changes.
During this change a bug was discovered that onEntry callback was
called before entering to regions of the state. This resulted the issue
that inside the onEntry a fire could change the active state of the
machine and then the regions of an inactive state would be initialized.
This was fixed by doing onEntry callback as the last step when entering
into a new state after entering to regions.