feat!: provide fire event and argument to presenter
This feature extends the function signature of the present member of
PagelessCreator. They are the event that caused the state transition
and the fire argument when that event was fired. This allows easily
passing information, like an error code, to the UI that in turn possibly
translates that error code to something human readable.