refactor!: replace present function with Presenter class
Original present function in PagelessCreator included parameters whose types
(the Machine and the event type) are defined inside the logic layer of the
application where state machines are typically used.
This means, as this present function is expected to be defined
in the UI layer, there would be an import from the UI (higher) layer to the
logic (lower) layer. In other words, internals of the logic layer would be exposed
to the UI layer.

In order to overcome this, the present function is replaced with the Presenter
abstract class that includes a present method that does not have any logic
layer specific parameter types. The Presenter abstract class needs to be implemented
to be used in PagelessCreator constructors. Any required logic layer data can be
added to this class as member variables and initialized in the constructor. This way
we are still able to provide data from logic layer to the present method.
Certainly, when doing so, in order to avoid importing from the logic layer one needs
to make sure the data added does not have types from the logic layer and when
needed apply translation of the data to types known to the UI layer.

Same approach is applied to the experimental SnackBarCreator class with the
SnackBarPresenter class.