fix: for bug of signaling UiClosed when UI was not closed
HismaRouterDelegate signals when initiates fire if the a UI component
(page) was already closed. This shall only happen when onPopPage
was called indicating that a page was already closed by a pop call.
The problem was that UiClosed was also signalled when the popRoute was
called which happens when the Android back button is pressed.
As this did not mean an already closed page, indicating UiClosed in
this case was a bug and resulted an assertion error in NavigationMachine
fire when Android back button was used:

`UI element was closed but the event defined in its creator led to
a state that is not present in the stack (the path is not forming a
circle). Check your mapping in your corresponding HismaRouterGenerator.`

This was fixed by distinguishing fire originated from popRoute
and onPopPage and only signal UiClosed when processing onPopPage.

Corresponding auto-tests were also added in t04_imperative_test.dart.