fix: double tap Android back button crash
If Android back button was tapped when HismaRouterDelegate state
stack included pages and the state machine arrived back to a state that
was in the stack (a circle was formed) the NavigatorMachine starts popping
those pages in the stack. If the Android back button was double tapped
it resulted `Bad state: Future already completed` exception to be thrown
from the invocation of route.didPop.
Fix avoid this unwanted processing of the same route multiple times by
saving the previous route and if the current route is the same as the
previous route processed by onPopPage then simply returns false.
Corresponding tests has been added to t04_imperative_test called
"Double tap of Android back button test, rootNavigator: false".